+++ .env.dev
@@ -1,10 +1,23 @@
+# RUN (docker container)
 PORT=3000
 NODE_ENV=development
-PG_USER=xxxx
-PG_PASSWORD=xxxx
+PG_USER=postgres
+PG_PASSWORD=****
 PG_PORT=5432
-PG_HOST=localhost
-PG_DATABASE=xxxx
-JWT_SECRET=xxxx
-JWT_EXPIRES_IN=xd
-JWT_COOKIE_EXPIRES_IN=x
+PG_HOST=postgres
+PG_DATABASE=****
+JWT_SECRET=****
+JWT_EXPIRES_IN=*d
+JWT_COOKIE_EXPIRES_IN=*
+
+# RUN (localhost)
+# PORT=3001
+# NODE_ENV=development
+# PG_USER=postgres
+# PG_PASSWORD=postgres@admin
+# PG_PORT=5430
+# PG_HOST=localhost
+# PG_DATABASE=hux
+# JWT_SECRET=mySecret
+# JWT_EXPIRES_IN=90d
+# JWT_COOKIE_EXPIRES_IN=90
