+++ src/interfaces/index.ts
@@ -1,2 +1,3 @@
 export * from './response';
 export * from './environment';
+export * from './type';
