+++ src/interfaces/response.ts
@@ -7,6 +7,7 @@ export enum RequestPath {
   BODY = 'body',
   QUERY = 'query',
   PARAMS = 'params',
+  HEADERS = 'headers',
 }
 
 type AppResponse = {
@@ -39,9 +40,6 @@ interface FailureResponse {
   };
 }
 
-export interface ObjectLiteral {
-  [props: string]: any;
-}
 
 type NotFoundError = {
   code: number;
@@ -53,30 +51,6 @@ type NotFoundError = {
 interface NotFoundResponse {
   error: NotFoundError;
 }
-
-type Config = {
-  APP: {
-    NAME: string;
-    VERSION: string;
-    VER: string;
-    DESCRIPTION: string;
-    AUTHORS: string;
-    HOST: string | undefined;
-    BASE_URL: string | undefined;
-    PORT: string | number;
-    ENV: string;
-    JWT_SECRET: string;
-    JWT_EXPIRES_IN: string;
-    JWT_COOKIE_EXPIRES_IN: string;
-  };
-  DB: {
-    USER: string;
-    PASSWORD: string;
-    PG_PORT: string | number;
-    HOST: string;
-    DATABASE: string;
-  };
-};
 
 type IUser = {
   id: string;
@@ -97,6 +71,5 @@ export {
   SuccessResponse,
   FailureResponse,
   NotFoundResponse,
-  Config,
   IUserResponse,
 };
