+++ src/interfaces/type.ts
@@ -1,0 +1,11 @@
+enum Role {
+  USER = 'user',
+  ADMIN = 'admin',
+}
+
+interface ObjectLiteral {
+  [props: string]: any;
+}
+
+
+export { Role, ObjectLiteral };
